+++ contrib/scripts/cover.sh
@@ -21,12 +21,12 @@ set -e
 # create coverage output
 echo 'mode: atomic' > $OUT
 for PKG in $(go list ./...|grep -v -E 'vendor|contrib|wiki|customtok'); do
-  if [[ "$TRAVIS_BRANCH" =~ master|release\/ && $TRAVIS_OS_NAME == "linux" ]]; then
-    echo "Running tests with race flag."
-    go test -v -race -covermode=atomic -coverprofile=$TMP $PKG
-  else
+  #if [[ "$TRAVIS_BRANCH" =~ master|release\/ && $TRAVIS_OS_NAME == "linux" ]]; then
+    #echo "Running tests with race flag."
+    #go test -v -race -timeout=30m -covermode=atomic -coverprofile=$TMP $PKG
+  #else
     go test -v -covermode=atomic -coverprofile=$TMP $PKG
-  fi
+  #fi
   tail -n +2 $TMP >> $OUT
 done